Separated ProductQuantity components out and removed duplicate code.
