+++ src/components/body/Checkout.js
@@ -9,6 +9,7 @@ import {ReactComponent as MasterCard} from '../../assets/icons8-mastercard.svg';
 import {ReactComponent as Discover} from '../../assets/icons8-discover.svg';
 import {ReactComponent as AmericanExpress} from '../../assets/icons8-amex.svg';
 import calculateTotal from '../../helpers'
+import ProductQuantity from './ProductQuantity'
 
 function ShippingAddress({user}) {
     return (<div className="checkout-part">
@@ -58,153 +59,6 @@ function Review() {
     return (<div className="checkout-part">
         <span className="checkout-list">3</span> <span className="checkout-list-steps">Review items and shipping</span>
     </div>)
-}
-
-function CartClearButton({product}) {
-    const dispatch = useDispatch()
-
-    let products = useSelector(state => state.products)
-    let total_items = useSelector(state => state.total_items)
-    let product_counter = useSelector(state => state.product_counter)
-
-    async function clearFromCart(product) {
-        products[product.id].pop()
-        product_counter[product.id] -= 1
-        delete products[product.id]
-        delete product_counter[product.id]
-
-        dispatch({
-          type: set_products,
-          products: products
-      })
-        dispatch({
-          type: set_product_counter,
-          product_counter: product_counter
-      })
-        dispatch({
-          type: set_total_items,
-          total_items: --total_items
-      })
-        dispatch({
-          type: set_total_cost,
-          total_cost: await calculateTotal(products)
-        })
-    }
-
-    return (
-        <span className="cart-product-quantity-delete" onClick={() => clearFromCart(product)}>Delete</span>
-    )
-}
-
-function ProductQuantityDecrementer({product, quantity}) {
-    const dispatch = useDispatch()
-
-    let products = useSelector(state => state.products)
-    let total_items = useSelector(state => state.total_items)
-    let product_counter = useSelector(state => state.product_counter)
-
-    async function removeFromCart(product) {
-        dispatch({
-          type: set_total_items,
-          total_items: --total_items
-        })
-        decrementCart(product)
-        dispatch({
-          type: set_total_cost,
-          total_cost: await calculateTotal(products)
-        })
-    }
-    async function decrementCart(product) {
-        if(!(product.id in products)) {
-          product_counter[product.id] = 0
-        } else {
-            products[product.id].pop()
-            product_counter[product.id] -= 1
-        }
-        dispatch({
-          type: set_products,
-          products: products
-      })
-        dispatch({
-          type: set_product_counter,
-          product_counter: product_counter
-      })
-    }
-
-    return (
-        <div className="cart-container-product-quantity-decrement">
-                  {
-                    (quantity < 2) ?
-                      <span className="cart-product-quantity-decrement"> - </span>
-                      :
-                      <span className="cart-product-quantity-decrement" onClick={() => removeFromCart(product)} disabled={quantity === 1}> - </span>
-                  }
-                </div>
-    )
-}
-
-function ProductQuantityIncrementer({product}) {
-    const dispatch = useDispatch()
-
-    let products = useSelector(state => state.products)
-    let total_items = useSelector(state => state.total_items)
-    let product_counter = useSelector(state => state.product_counter)
-
-    async function addToCart(product) {
-        dispatch({
-          type: set_total_items,
-          total_items: ++total_items
-        })
-        incrementCart(product)
-        dispatch({
-          type: set_total_cost,
-          total_cost: await calculateTotal(products)
-        })
-    }
-    async function incrementCart(product) {
-        if(!(product.id in products)) {
-          products[product.id] = [product]
-          product_counter[product.id] = 1
-        } else {
-            products[product.id].push(product)
-            product_counter[product.id] += 1
-        }
-        dispatch({
-          type: set_products,
-          products: products
-      })
-        dispatch({
-          type: set_product_counter,
-          product_counter: product_counter
-      })
-    }
-
-    return (
-        <div className="cart-container-product-quantity-increment">
-                  <span className="cart-product-quantity-increment" onClick={() => addToCart(product)}>
-                      +
-                  </span>
-                </div>
-    )
-
-}
-
-function ProductQuantity({product, quantity}) {
-    return (
-        <div className="cart-container-product-quantity">
-              <div className="cart-product-quantity">
-                <ProductQuantityDecrementer product={product} />
-                <div className="cart-container-product-quantity-label">
-                  <span className="cart-product-quantity-label">
-                    {quantity}
-                  </span>
-                </div>
-                <ProductQuantityIncrementer product={product} />
-              </div>
-              <hr className="cart-product-quantity-divider"/>
-              <CartClearButton />
-            </div>
-    )
 }
 
 function ProductStock({product}) {
